more tests pass (but not all)

+++ src/concat.c
@@ -66,7 +66,8 @@ char *append_to(char *s1, char *s2)
 char *str_splice(char *src, char *ins, int pos, int len)
 {
 	assert (pos >= 0);
-	assert (pos + len <= (unsigned int) strlen(src));
+	// TODO: use unsigned ints, so gcc stops flagging error
+	assert (pos + len <=  strlen(src));
 
 	int insert_len = strlen(ins);
 	int result_len = strlen(src) + insert_len - len;
